+++ lib/services/local_note_service.dart
@@ -24,7 +24,8 @@ class LocalNoteService {
     return List.generate(maps.length, (i) => Note.fromJson(maps[i]));
   }
 
-  Future<Note?> getNoteById(int id) async {
+  Future<Note?> getNoteById(double id) async {
+    // double
     final db = await _dbHelper.database;
     final List<Map<String, dynamic>> maps = await db.query(
       'notes',
@@ -56,7 +57,8 @@ class LocalNoteService {
     );
   }
 
-  Future<void> deleteNote(int id) async {
+  Future<void> deleteNote(double id) async {
+    // double
     final db = await _dbHelper.database;
     await db.delete(
       'notes',
